refactor generating models
